:truck: Move mssql simple types into db/utils folder.
Why? To reduce clutter.
