+++ api/src/db/templates/sample-migration.ts
@@ -1,7 +1,7 @@
 import { DataTypes } from "sequelize"
 
 import type { Migration } from "@/db/umzug"
-import { MssqlSimpleTypes } from "@/db/mssql-simple-types"
+import { MssqlSimpleTypes } from "@/db/utils/mssql-simple-types"
 
 export const up: Migration = async ({ context: queryInterface }) => {
   throw new Error("Not implemented")
