+++ api/src/db/utils/mssql-simple-types.ts
